+++ cmd/finish/main.go
@@ -133,14 +133,24 @@ func main() {
 			}
 		}
 
+		am, err := pipelinectxt.ReadArtifactsManifestFromFile(
+			filepath.Join(pipelinectxt.ArtifactsPath, pipelinectxt.ArtifactsManifestFilename),
+		)
+		if err != nil {
+			log.Fatal(err)
+		}
 		for artifactsSubDir, files := range artifactsMap {
 			for _, filename := range files {
-				nexusGroup := fmt.Sprintf("/%s/%s/%s/%s", ctxt.Project, ctxt.Repository, ctxt.GitCommitSHA, artifactsSubDir)
-				localFile := filepath.Join(pipelinectxt.ArtifactsPath, artifactsSubDir, filename)
-				fmt.Printf("Uploading %s to Nexus repository %s, group %s ...\n", localFile, nexusRepository, nexusGroup)
-				err = nexusClient.Upload(nexusRepository, nexusGroup, localFile)
-				if err != nil {
-					log.Fatal(err)
+				if am.Contains(nexusRepository, artifactsSubDir, filename) {
+					log.Printf("Artifact %s is already present in Nexus repository %s.", filename, nexusRepository)
+				} else {
+					nexusGroup := nexus.ArtifactGroup(ctxt, artifactsSubDir)
+					localFile := filepath.Join(pipelinectxt.ArtifactsPath, artifactsSubDir, filename)
+					fmt.Printf("Uploading %s to Nexus repository %s, group %s ...\n", localFile, nexusRepository, nexusGroup)
+					err = nexusClient.Upload(nexusRepository, nexusGroup, localFile)
+					if err != nil {
+						log.Fatal(err)
+					}
 				}
 			}
 		}