+++ CHANGELOG.md
@@ -14,6 +14,7 @@ listed in the changelog.
 - Generating a SonarQube report fails when PR exists for scanned branch ([#227](https://github.com/opendevstack/ods-pipeline/issues/227))
 - Generating a SonarQube report fails when background task does not finish immediately ([#227](https://github.com/opendevstack/ods-pipeline/issues/227))
 - Quality Gate check fails due to incorrect API authentication (detected while working on #227)
+- Uploading of artifacts in `ods-finish` may fail when artifact is already present from previous pipeline run ([#255](https://github.com/opendevstack/ods-pipeline/issues/255))
 
 ### Changed